requirements.txt
streamlit==1.28.1
torch==2.0.1
transformers==4.34.0
pandas==2.1.1
plotly==5.17.0
requests==2.31.0
scikit-learn==1.3.0
wordcloud==1.9.2
python-dotenv==1.0.0
geopy==2.4.0
folium==0.14.0
streamlit-folium==0.15.0
emoji==2.8.0
numpy==1.24.4
textblob==0.17.1
Pillow==10.0.0
nltk==3.8.1

